+++ _src/tests/browser-client/src/app/app.component.ts
@@ -7,6 +7,11 @@ import { NaniumJsonSerializer } from '../../../../serializers/json';
 import { NaniumConsumerBrowserHttp } from '../../../../managers/consumers/browserHttp';
 import { TestClientRequestInterceptor } from '../../../interceptors/client/test.request.interceptor';
 import { Nanium } from '../../../../core';
+import { NaniumProviderBrowser } from '../../../../managers/providers/browser';
+import { ServiceRequestBase } from '../../../services/serviceRequestBase';
+import { ClientServiceExecutionContext } from '../services/clientServiceExecutionContext';
+import { StuffGetExecutor } from '../services/stuff/get.executor';
+import { StuffGetRequest } from '../services/stuff/get.contract';
 
 @Component({
 	selector: 'app-root',
@@ -16,6 +21,21 @@ import { Nanium } from '../../../../core';
 export class AppComponent implements OnInit {
 	testGetResponse?: ServiceResponseBase<TestGetResponseBody>;
 	error?: any;
+
+	browserProvider = new NaniumProviderBrowser({
+		isResponsible: async (request, serviceName) => {
+			return serviceName.startsWith('NaniumClientTest:') ? 2 : 0;
+		},
+		isResponsibleForEvent: async (eventName) => {
+			return eventName.startsWith('NaniumClientTest:') ? 2 : 0;
+		},
+		requestInterceptors: [new class {
+			async execute(request: ServiceRequestBase<any, any>, context: ClientServiceExecutionContext): Promise<ServiceRequestBase<any, any>> {
+				context.user = { id: 1, name: 'TestUser' };
+				return request;
+			}
+		}]
+	});
 
 	async ngOnInit() {
 		this.initNanium();
@@ -31,9 +51,16 @@ export class AppComponent implements OnInit {
 			requestInterceptors: [TestClientRequestInterceptor],
 			handleError: async (err: any): Promise<any> => {
 				throw { handleError: err };
-			}
+			},
+			isResponsible: async (request, serviceName) => {
+				return serviceName.startsWith('NaniumTest:') ? 2 : 0;
+			},
+			isResponsibleForEvent: async (eventName) => {
+				return eventName.startsWith('NaniumTest:') ? 2 : 0;
+			},
 		});
 		Nanium.addManager(naniumConsumer).then();
+		Nanium.addManager(this.browserProvider).then();
 	}
 
 	async test1(): Promise<void> {
@@ -48,8 +75,8 @@ export class AppComponent implements OnInit {
 		try {
 			const request = new TestBufferRequest({
 				id: '1',
-				buffer1: new NaniumBuffer('123'),
-				buffer2: new NaniumBuffer('456')
+				buffer1: new NaniumBuffer(new TextEncoder().encode('123')),
+				buffer2: new NaniumBuffer(new TextEncoder().encode('456'))
 			});
 			const response = await request.execute();
 			console.log(response.id === '1');
@@ -62,6 +89,15 @@ export class AppComponent implements OnInit {
 		}
 	}
 
+	async test3(): Promise<void> {
+		try {
+			this.browserProvider.addService(StuffGetRequest, StuffGetExecutor);
+			const result = await new StuffGetRequest().execute();
+			console.log('value: ', result.value, ' ?=== "TestUser"');
+		} catch (e) {
+			this.error = e;
+		}
+	}
 
 	// execute request via the consumer